test(transaction): add command integration tests
